Feat: added ingame settings
