+++ client/Assets/Scenes/Menu.unity
@@ -1996,7 +1996,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 724175318}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 6.73999, y: 0, z: -10}
+  m_LocalPosition: {x: 6.73999, y: 0, z: 0}
   m_LocalScale: {x: 1.9124, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
