+++ client/Assets/Scripts/UI Scripts/BGScroll.cs.meta
