+++ client/Assets/Scripts/UI Scripts/PauseManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 71699341ac190cc4780d25a245eb9df6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
