+++ client/Assets/VFX/UI-Assets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0059100ff6cd13141b4ef03dbcec541c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
