+++ client/Assets/VFX/UI-Assets/Wood-UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f3f00cb1814d42347b7f4d34ea0dcaa8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
