+++ client/Logs/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,0 +1,3 @@
+Base path: 'D:/Unity/Unity Hub/2021.3.23f1/Editor/Data', plugins path 'D:/Unity/Unity Hub/2021.3.23f1/Editor/Data/PlaybackEngines'
+Cmd: initializeCompiler
+
